test: complexity enforcement
